third_party/blink/web_tests/fast/events/event-isolated-world-clone-expected.txt

CONSOLE MESSAGE: CustomEvent received in main world
CONSOLE MESSAGE: detail was {"foo":5,"bar":"hello","targetWorld":"main"}
CONSOLE MESSAGE: CustomEvent received in isolated world
CONSOLE MESSAGE: detail was {"foo":5,"bar":"hello","targetWorld":"isolated"}
CONSOLE MESSAGE: MessageEvent received in main world
CONSOLE MESSAGE: data was {"foo":5,"bar":"hello","targetWorld":"main"}
CONSOLE MESSAGE: MessageEvent received in isolated world
CONSOLE MESSAGE: data was {"foo":5,"bar":"hello","targetWorld":"isolated"}
CONSOLE MESSAGE: PopStateEvent received in main world
CONSOLE MESSAGE: state was {"foo":5,"bar":"hello","targetWorld":"main"}
CONSOLE MESSAGE: PopStateEvent received in isolated world
CONSOLE MESSAGE: state was {"foo":5,"bar":"hello","targetWorld":"isolated"}
Tests that properties of various events are cloned between isolated worlds.
